YT-20886: Unify the interface type for the table index 5b41b3937461497b80006bfde33558f7a152bc0f

+++ yt/yt/ytlib/api/native/partition_tables.h
@@ -60,7 +60,7 @@ private:
     void AddUnversionedDataSlices(const TInputTable& inputTable);
     NTableClient::TComparator GetComparator(int tableIndex);
     void FixLimitsInOrderedDynamicStore(
-        size_t tableIndex,
+        int tableIndex,
         const std::vector<NChunkClient::TInputChunkPtr>& inputChunks);
     void ValidatePaths();
 };
